Add files via upload
Secondlife Logs from the Group Got Roleplay. Where again SamanthaPrater AKA Emily Cassidy tried to change the rules and how a group was run simply because the owner was not around to stop her.  First instant where I was privy to her taking advantage of the fact, the owner was not around to stop her.