Add files via upload
And into the new year she is still screaming about James.

So the latest bullshit is her screaming James is some Emily Rose scammer. Yeah. Emily Rose is an Actress that was on Haven the series that aired on scifi. Why is Emily attached to this. Because she found some odd ball shit in an email from secondlife of someone called Emily Rose. Likely a simple wrong IM, sent it to the wrong person and because they never heard from the person, realized they messaged wrong person and got in touch with the right one. But to Emily that is all some conspiracy that James, her step dad, scammed people.  

And also look at the songs. Is that creepy? Its creepy. She is associating songs meant for an ex lover to her Step Dad. As I was telling my Cohort in crime AKA Ruin, lastnight. I feel that due to the schizoaffective disorder, she is trying to put herself into her mothers place, so James will react to her the in the same way, IE loving and nurturing. Of course he will not, because she is not her mother and she has tormented him for ages now. I really do believe it is because she is infatuated with james being the older father figure and... BLEGH! Its not a pretty sight to say the least.