Add files via upload
More James, EmilyRoseScam screaming. There is no scam, its her delusional mind making up stuff to blame her Step Dad. And Only Fans. Lords help us. And her pricing herself on Instagram. Well that is a reportable offense. And we will be sending this off to the disability administration. Sucks to be her.