Add files via upload
Part 2 of her screaming. Long Short, More james. Real Quick. None of us are mad at him, but we all do not like her. She screams about her timeline, but its all skewed, we know it is, we have spoken to others and we know what is true and not. Anything she is saying is not true.

Findom is sex work. Dominating someone gets their sexual needs met, they masturbate(Self Service Sex) and you lather rinse repeat. That is why SL is so appealing.

Again no one is mad at James, Findom is sex work, and she is a delusional nutcase