Add files via upload
Here we go again. More James. Look if you read through this, James's words are in here. He doesn't bring her up unless she out right attacks his home and life. Most cases, we do not hear from him unless she sends the cops at him, or other people. Which she has done, she screams she doesn't but, she has. And no one in the Dibney Cru believes her lies. Because again, it has changed. Her story has changed again. It always does. She is back to screaming James knew her at 9. Lisa nixed that quickly and said no James didn't come into their lives till she was into her later years like 19teen 20 years old. So no one is believing her BS.

Hacked. Lord help us, we know she was not hacked. We have a lot of the proof that WillowtheWay or whatever the Willow name is, was not hacked. Because we found all of it. So again, its a load of BS. She does this, as one of the girls said in the discord, that she screams hacked when she is caught being abusive. Which is every day. 

We all know she is crazy but I want to bring up something I find hysterical. So she keeps bringing up her IQ. She has admitted that she had it done in her late teens and it was a Mensa score of 77... Yeah. At the height of when those tests are often taken, she scored a 77. Which makes her a moron. An idiot. Average is 85 to 95-120 depending on who you take the test through.  So again she is a moron. We are not calling her an Idiot because its an opinion, we are calling her an idiot because it is fact. Comparatively against many of us in the discord we run IQ scores from 95 to 162.  Mine is 132 which to give you an idea this is what that means 

"A score of 116 or more is considered above average. A score of 130 or higher signals a high IQ. Membership in Mensa, the High IQ society, includes people who score in the top 2 percent, which is usually 132 or higher." Jan 28, 2020 

Her score of 77 puts her here "70 to 84: Borderline mental disability." which we know. Also "Borderline intellectual functioning, also called borderline mental retardation (in the ICD-8), is a categorization of intelligence wherein a person has below average cognitive ability (generally an IQ of 70–85), but the deficit is not as severe as intellectual disability (below 70)." So not retarded, but not smart either.

We know for a fact, she googles everything, and often does not comprehend all she reads, twisting it to fit her narrative. We have seen it time and time again. I have seen it when she attacked me before the Dibney Cru was formed.  It is like a synapse is not firing right to allow her to comprehend what she is reading does not always apply to her or can be used by her. Take law for instance.

But I am rattling. Anyways. That is the past 24 hours.