Add files via upload
The Weekend of Jan 15th to Monday the 18th

Holy Crazy ass shit! So this weekend has been again Emily screaming about James, pretty normal. But we had a twist in the events on Sunday. She started to attack Elk_Administration on Reddit, Facebook, and anywhere else she could possibly slither her way into. The problem is the Dibney Cru jumped in. And it has been hours of her screaming and blaming everyone for everything she has done.
She also went back after Candy. Which to her dismay we jumped in again and nixed that. Cause Candy is with us, and it took 2 seconds for us to post the images, and alert her Emily was back to her BS about her.