Add files via upload
Jaaaammeeeesssss! For those that are new here. James is not in the Discord for the Dibney Cru, he never has been.  We have felt that it would do more harm to him and to us to have him in it. Not because James is a danger or anything, but honestly, he has enough on his plate taking care of Ronda and doing his robotics. 

She keeps screaming about her proof. Lordy. Her proof is spewing and hissing at her screen like the idiot she is. Frankly I love her reactions with Elk. Threat after threat.  As Ruin said  "A threat is something you make when you are not sure you can pull it off." Or its a way to posturing and preening and means nothing. None of us take what she says as serious. 

She is still after Candy. Despite her words of late, we know better. And it will be interesting to see when this is done.