Add files via upload
Oh last night was hilarious! So she got a hold of this lovely repo. And the first thing she attacks is the Aliases. Oh I about died on that. Because she obviously does not know what Alias means: Thank you Webster for this.

"An alias name or AKA (also known as) is any name that has been used by the candidate in the past. Life events such as marriage and divorce result in many candidates with records associated with more than one name. It is possible that criminal records could exist under any prior name." 

In other words, that list is a known list of all her Aliases or AKA's. So while she scribbled all over it(btw breaking the copyright I have on this repo, its on the front page), she proved that we were right on all those names. Sucks to be here.

And in the midst of all this, she goes off on me and this repository. To clarify something. This repository was originally created as a means to see all the harassment coming from her via Notecards, and FLICKR messages, and anything else she wanted to throw at me, so I could see for myself that I was not imagining anything. It has turned into a place to dump all the crap she spews for when we finally need it. But I am not the only one who has started a repository. Squirrel(lovely sarcastic brilliant Squirrel) has done the same.  Its not paranoia or I need a hobby.

I do this for 10 minutes out of every morning. Thats it. The information that often gets thrown in here now, comes from screen grabbers in my discord. There are a lot of us that contribute. So when I say this is a work of love, I just organize things into neat little folders, those girls and guys that hate her guts, do all the work. I have a job, a family, and hobbies, like finding patterns right now for kilts for my husband and son.

Lets touch on that too. She likes to focus on 1 child. My daughter. I have 2. My daughter right now is in that teenage state of only wanting mom when she needs an opinion about her drawing, nibbles, or cuddles(she acts like a cat and I am not against this). My Son, is still a little sprout. And he takes up a lot more time than my daughter does. Not because I favor him more, she doesn't need me right now as much as he does, and she will tell me when she needs me. So while I find it funny Emily only focuses on my daughter, I have 2 children. 

Speaking of my daughter, and this has come up in the past, many of us feel it is because Cheshire is Autistic, that Emily focuses so hard on her and not my son(who has a learning disability as well). She tries to associate herself with Cheshire not realizing, Cheshire and her are NOTHING alike and using Cheshire as a means to get me on her(Emily's) side only does the opposite.  Cheshire is a beautiful, loving, empathetic teenager with immense amount of raw artistic talent that we have been helping focus her with for ages. She is smarter than Emily will ever be, kinder, doesn't bully people, not even her brother, and knows how to treat people. She understands the words no, and leave me alone, stop, and not right now. She does not push anything on anyone, but will make a pretty decent argument when she has to. She doesn't need to spam me or her father to get something she wants, she simply asks the right way the First TIme! Unlike Emily who is a complete Nightmare of a person.

IQ! HOLY HELLS I woke up to this and damn near laughed my ass off. So she tried to say that her chromosomes can not lead to her having a low IQ. Lets clarify this, I am bringing in the words of the wise person I know who knows all about this.

"She has a microdeletion of a small portion of a chromosome. I also already explained that to her long ago when she was whining in a voice chat about being missing chromosomes and I explained exactly what 22Q is and what tiny fraction of a chromosome is damaged. If she was missing an entire chromosome, much less several like she says, she wouldn't have been born alive. And IQ isn't compared to chromosomal count at all. Maybe because of the Q in 22q so...her usual being a moron and not able to read properly." (This conversation can be found above in today's commit)

She has also gone off recently, about having her IQ tested by her doctors when she was a kid, hence, they took into consideration her mental incapability's at the time of the test. Unless she did one online, then the odds were already stacked against her. And its been stated by several renowned websites and professionals, that IQ deteriorates over age.  So her age has likely lowered that number as well further. The damn thing about this is, its not her IQ that we base a lot of our Idiot comments off of, frankly its her actions. 

And the last 4 images is indicator she is in the apology phase again. I think this is 3 times in the past 2 weeks. Honestly we have in the Dibney Cru simply laughed it off. Its not sincere she doesn't mean it, and we have heard it and seen it so many times, its just words.  Now if she actually stopped her drama mongering, shut up about everyone and I mean everyone, this whole thing would have died ages ago. But she simply can not. Again that leads back to the Narcissism that is prevalent in her actions. And her own words for that matter. She needs this attention.

Oh a quick things: We found her tiktok thanks to her blasting that when she doodled on the aliases page, if I am not mistaken, most of us on TikTok now have her blocked. I know I do. 

Have a good day folks!