Update
Let me start by saying, go watch this!

https://youtu.be/CIPydwez1Vs 

This man is hilarious. The Dibney Cru was in full swing and we had a blast talking. But besides that, we found out that the whole Emily Rose Scam the Trollorist was screaming about was done by her own hands. To make this a bit short. She used the SammanthaPrater@gmail.com Email for the Emily Rose Craigslist searches and replies she was doing to get booty calls. She uses that same email to harass Gaea, and anyone else for that matter. I have a few emails from that email address of hers myself. But she emailed James using that Email. Yup you read that right. Don't believe me, or do believe me for that matter, just look at the images here, I promise you won't be disappointed!

Also quick note: This Github has been backed up on 3 other sites, so if Emily is reading this and thinks she can make this disappear, sorry turd, its not going anywhere. It is also backed up on a flash drive, an External, my old pc, and several other sources. And all of them get updated on the regular.