IP Lies and Emails
So yesterday was a lesson in idiocy. She attacked James's youtube with accusations that the Phones he has are not hers. I will be uploading the voice mails proving that those phones are hers. One in particular has her screaming "LEAVE MY PHONES ALONE!" As James says all the time "I just shit myself" because when I heard it, all doubt was removed. It will be a 2 part upload as James has over 218 voice mails from this psycho. 

Another thing she attacked him over was the IP address. Okay I have included in this commit, how Static IP works. Layman's terms: It is the permanent address on the phone itself. It is traceable. It is also possible for a phone to have up to 6 different IP address, but the static IP does not change. Also too sim cards are registered to a person, further identifying the device. So he has 2 phones with sim cards with her name all over it. And with static IP addresses connected to her. Funny how when you look up how things actually work, you find out the actual truth instead of cherry picking to fit your narrative, huh!

The ADA. She has been screaming that we are "Picking on her" over her disabilities. No. Just cause she was in Spec Ed does not mean we are picking on her. And the ADA does not cover internet dealings. What it does cover is discrimination in the work and business places. So for an example a work place can not discriminate against an individual with a disability. Same goes for business dealings. For example, if someone with a disability was trying to own a home, the brokerage could not discriminate against them. What it does not cover is trolling and internet dealings. I have disclosed the PDF file and the screen grab of that in this commit.

As well as these words. There are those of us who have been in SPEC ED, or have Family that are in Spec Ed. We know the laws and the workings of the disability acts because we live with them or advocates for them. What we are pointing out is her Actions, her constant harassment, lies, and other insane ramblings. As well as her Cyber Attacks on individuals if they chose to could press charges against her for said harassment and slander. And included in that the threats made against their physical lives, businesses, hobbies, and possible relations.

And the Grand FInale! So she posted her email lastnight. So I put it above the collab that was made showing that the phones james has, the email she recently sent him, and the blog post all show the SAME EMAIL ADDRESS. Yup. All 3 have the same Email address. This is so much better than a smoking gun.  Because not only did she confirm publicly that the email is hers, and you can tell its hers in all three images, but the fact that she can not decipher the truth of what was shown. Those emails have nothing to do with IP addresses or anything else like she claims, like hacking. It confirms she was using the Alias of Emily Rose to try and solicit sex/prostitution on craigslist. She was in fact using that name. IT confirms she has used that email to target her mom's boyfriend with hate.  It confirms that it is her that has done all this horrible acts. 

Evidence is in this repository and its backups.