Reddit and the blissful 36 hours of silence
A lot of this is Reddit. Lords she screamed. 

But the blissful part was we had 36 hours of complete silence. Yes, yes you read that right, 36 complete hours of silence! And we had 17 hours yesterday, till the howler beast woke up. 

What no one seems to understand is we look forward to those silences. We don't do any of this for fun. Eventually this will be used to help those victims that she has harmed. Meantime, not a single one of us does this for fun. Personally I would rather be starting half the patterns that Aislinn has sent me, or working on projects that I have left to the wayside because of detail organizing this. I would rather be working on my schooling or anything else.