Aevum DIscord Server Pt2
I had to look up the name and found out she is stealing this name from several people and companies. The biggest is an aeronautics and robotics company that launched a completely autonomous craft named Ravn X. Its worth is more than she will ever see in her life and likely the next five lifetimes.
Its a name of a band.
And its the name of a Study in Medieval Studies and Culture. 
One gentleman uses it to name is Course at Cornell University in New York.

So she stole another name. No shocker there.