Michigan Wire Tapping
Much like her misunderstanding of Michigan laws. We are still a 1 party consent state. It has not been legally certified by MI legislature or the Sixth District Court that we are not(circa Oct 2020). But that only pertains to Wire Tapping and face to face conversations. That does not include internet conversations or collecting of publicly(not public ally) posted documents or information. 

That also does not include voice mails. See what miss Emily doesn't understand is that once you send a message or make a voice mail to someone, it then becomes that persons property. You consented to their use by making the message and hitting send. If you do not want someone to have such items,do not contact them, send messages, or create voice mails.

Also do not post on social media anything that can be used against you later on, in any way shape or form.