Valentines Day Weekend
She went back on her treaty. Has been back to spam emailing James. Using Cholo and trying to justify the derogatory term and so much more.

Personally thinking about it. She is lonely. She has no one to talk to, and when James, Aislinn and Vixen got in the call with her, she had human contact. As soon as that went away, she was back up to her old games. Its disappointing.