Aww hating on the Repo!
The Two images up there says most of this, but frankly I don't give a fuck if she hates on this. Like I said there are copies of this all over, not including what I have currently sent to James, via a flash drive of her bullshit. Which included: 
All of the background checks several of us have done
All the Pedo shit, both her talk of him, and the actual files from what could be found via a google search, and everything about him she spewed from the old disney forums. Which mind you is what brought our attention upon that. Her screams of background checks and such, led us right to that sicko. 
Her hallucinations that were occurring on the old Disney Forums, IE Royal Hitler, her relation to Cleopatra and other rants. And so much more. 

And she seems to forget there is Squirrels Repo as well that has more sitting there, and the Trollorist Forums, as well as several blogs. 

But her hate on this is to get MY Attention. And the most it has got was a chuckle and me going back to playing Pokemon Go with my actual go outside with them and hang out friends! And well this little blurb.